Refactor prop organization
